Stop `init.sh` from duplicating anime-subs link

+++ init.sh
@@ -3,7 +3,7 @@
 if [[ -e "./init.sh" ]]; then
 
     export PATH="$(pwd)/ingestion:${PATH}";
-    ln -sf "$IKNOW_DIR/anime-subs" ./anime-subs
+    [[ -e ./anime-subs ]] || ln -sf "$IKNOW_DIR/anime-subs" ./anime-subs
 
 else
     echo "Error: Must be sourced in $ERAB_DIR";
